requirements.txt
dateparser==1.2.0
pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
spotipy==2.23.0
schedule==1.2.1